+++ examples/okta_profile_mapping/prevent_delete.tf
@@ -1,5 +1,6 @@
 resource okta_profile_mapping test {
   source_id          = "${okta_idp_social.google.id}"
+  target_id          = "${data.okta_user_profile_mapping_source.user.id}"
   delete_when_absent = false
 
   mappings {
@@ -63,3 +64,5 @@ resource okta_idp_social google {
   client_secret     = "abcd123"
   username_template = "idpuser.email"
 }
+
+data okta_user_profile_mapping_source user {}