+++ sdk/mappings.go
@@ -37,8 +37,8 @@ func (m *ApiSupplement) RemovePropertyMapping(mappingId, id string) (*okta.Respo
 	return m.RequestExecutor.Do(req, nil)
 }
 
-func (m *ApiSupplement) GetProfileMappingBySourceId(sourceId string) (*Mapping, *okta.Response, error) {
-	url := fmt.Sprintf("/api/v1/mappings?sourceId=%s", sourceId)
+func (m *ApiSupplement) GetProfileMappingBySourceId(sourceId, targetId string) (*Mapping, *okta.Response, error) {
+	url := fmt.Sprintf("/api/v1/mappings?sourceId=%s&targetId=%s", sourceId, targetId)
 	req, err := m.RequestExecutor.NewRequest("GET", url, nil)
 
 	if err != nil {
@@ -100,4 +100,38 @@ func (m *ApiSupplement) UpdateMapping(mappingId string, body Mapping, qp *query.
 		return nil, resp, err
 	}
 	return &mapping, resp, nil
+}
+
+// FindProfileMappingSource retrieves profile mapping source/target via name
+func (m *ApiSupplement) FindProfileMappingSource(name, typ string, qp *query.Params) (*MappingSource, *okta.Response, error) {
+	uri := "/api/v1/mappings"
+
+	if qp != nil {
+		uri += qp.String()
+	}
+
+	req, err := m.RequestExecutor.NewRequest("GET", uri, nil)
+
+	if err != nil {
+		return nil, nil, err
+	}
+
+	mappings := []*Mapping{}
+	res, err := m.RequestExecutor.Do(req, &mappings)
+
+	for _, m := range mappings {
+		if m.Target.Name == name && m.Target.Type == typ {
+			return m.Target, res, nil
+		} else if m.Source.Name == name && m.Source.Type == typ {
+			return m.Source, res, nil
+		}
+	}
+
+	if after := GetAfterParam(res); after != "" {
+		qp.After = after
+
+		return m.FindProfileMappingSource(name, typ, qp)
+	}
+
+	return nil, res, fmt.Errorf("could not locate profile mapping source with name %s", name)
 }